commit for the morning

+++ src/main/java/com/techelevator/model/Conversion.java
@@ -9,11 +9,11 @@ public class Conversion {
 	private static double temperature;
 	
 	
-	public static double convertFtoC(int f){
+	public static double convertFtoC(double f){
 		temperature = ((f-32)*0.556);
 		return temperature;
 	}
-	public static double convertCtoF(int c){
+	public static double convertCtoF(double c){
 		temperature = ((c*1.8)+32);
 		return temperature;
 	}
